refactor: replace QueryFilters by ProductQueryFiltersServiceProvider
