+++ plugins/woocommerce/src/Blocks/BlockTypes/ProductFilters.php
@@ -1,8 +1,7 @@
 <?php
 namespace Automattic\WooCommerce\Blocks\BlockTypes;
 
-use Automattic\WooCommerce\Blocks\QueryFilters;
-use Automattic\WooCommerce\Blocks\Package;
+use Automattic\WooCommerce\Internal\ProductQueryFilters\FilterData;
 
 /**
  * Product Filters Block.
@@ -169,7 +168,7 @@ final class ProductFilters extends AbstractBlock {
 			'rating_counts'       => null,
 		);
 
-		$filters = Package::container()->get( QueryFilters::class );
+		$filters = wc_get_container()->get( FilterData::class );
 
 		if ( ! empty( $block->context['query'] ) && ! $block->context['query']['inherit'] ) {
 			$query_vars = build_query_vars_from_query_block( $block, 1 );
