+++ plugins/woocommerce/src/Container.php
@@ -30,6 +30,7 @@ use Automattic\WooCommerce\Internal\DependencyManagement\ServiceProviders\Restoc
 use Automattic\WooCommerce\Internal\DependencyManagement\ServiceProviders\UtilsClassesServiceProvider;
 use Automattic\WooCommerce\Internal\DependencyManagement\ServiceProviders\BatchProcessingServiceProvider;
 use Automattic\WooCommerce\Internal\DependencyManagement\ServiceProviders\LayoutTemplatesServiceProvider;
+use Automattic\WooCommerce\Internal\DependencyManagement\ServiceProviders\ProductQueryFiltersServiceProvider;
 
 /**
  * PSR11 compliant dependency injection container for WooCommerce.
@@ -79,6 +80,7 @@ final class Container {
 		LayoutTemplatesServiceProvider::class,
 		LoggingServiceProvider::class,
 		EnginesServiceProvider::class,
+		ProductQueryFiltersServiceProvider::class,
 	);
 
 	/**
