+++ plugins/woocommerce/src/Internal/DependencyManagement/ServiceProviders/ProductQueryFiltersServiceProvider.php
@@ -1,0 +1,33 @@
+<?php
+
+namespace Automattic\WooCommerce\Internal\DependencyManagement\ServiceProviders;
+
+use Automattic\WooCommerce\Internal\DependencyManagement\AbstractServiceProvider;
+use Automattic\WooCommerce\Internal\ProductQueryFilters\FilterData;
+use Automattic\WooCommerce\Internal\ProductQueryFilters\FilterClauses;
+
+/**
+ * LoggingServiceProvider class.
+ */
+class ProductQueryFiltersServiceProvider extends AbstractServiceProvider {
+	/**
+	 * List services provided by this class.
+	 *
+	 * @var string[]
+	 */
+	protected $provides = array(
+		FilterClauses::class,
+		FilterData::class,
+	);
+
+	/**
+	 * Registers services provided by this class.
+	 *
+	 * @return void
+	 */
+	public function register() {
+		$this->share( FilterClauses::class );
+
+		$this->share( FilterData::class )->addArgument( FilterClauses::class );
+	}
+}
